requirements.txt
python-dotenv
discord.py
discord.py-self